- "DDBJ" and "EMBL" are no longer valid refseq db_xrefs (e.g. NT_039169) - a little code cleanup (e.g. unused vars) - started fixing discrepancies in HTML mode.   - Since just starting this, there were actually many changes to be made,     even for basic records like U54469.   - minor code refactoring (just moved some shared information to     utils.hpp to avoid duplication ) JIRA GP-1267
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@49767 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/format/gather_items.cpp
@@ -1657,7 +1657,7 @@ static bool s_SeqLocEndsOnBioseq(const CSeq_loc& loc, CBioseqContext& ctx, EEnds
     CSeq_loc_CI last;
     for ( CSeq_loc_CI it(loc); it; ++it ) {
         if( mode == eEndsOnBioseqOpt_AnyPartOfSeqLoc ) {
-            if( seq.IsSynonym(it.GetSeq_id()) && it.GetRangeAsSeq_loc()->GetStop(eExtreme_Biological) < seq_len ) {
+            if( seq.IsSynonym(it.GetSeq_id()) && (int)it.GetRangeAsSeq_loc()->GetStop(eExtreme_Biological) < seq_len ) {
                 return true;
             }
         }
@@ -1674,7 +1674,7 @@ static bool s_SeqLocEndsOnBioseq(const CSeq_loc& loc, CBioseqContext& ctx, EEnds
     } else {
         if( endsOnThisBioseq ) {
             // if we're not partial, we also check that we're within range
-            return seq_len > last.GetRangeAsSeq_loc()->GetStop(eExtreme_Biological) ;
+            return seq_len > (int)last.GetRangeAsSeq_loc()->GetStop(eExtreme_Biological) ;
         } else {
             return true;
         }
